+++ Source/Mosa.Demo.CoolWorld.x86/Mosa.Application/ShowPCI.cs
@@ -3,6 +3,7 @@
 using Mosa.Demo.AppSystem;
 using Mosa.DeviceSystem;
 using Mosa.DeviceSystem.PCI;
+using Mosa.DeviceSystem.Service;
 
 namespace Mosa.Demo.Application
 {
