+++ Source/Mosa.Demo.CoolWorld.x86/Mosa.Application/Shutdown.cs
@@ -1,7 +1,7 @@
 ﻿// Copyright (c) MOSA Project. Licensed under the New BSD License.
 
 using Mosa.Demo.AppSystem;
-using Mosa.DeviceSystem;
+using Mosa.DeviceSystem.Service;
 
 namespace Mosa.Demo.Application
 {
