+++ Source/Mosa.DeviceSystem/Device.cs
@@ -1,5 +1,6 @@
 ﻿// Copyright (c) MOSA Project. Licensed under the New BSD License.
 
+using Mosa.DeviceSystem.Service;
 using System.Collections.Generic;
 
 namespace Mosa.DeviceSystem
