+++ Source/Mosa.DeviceSystem/Service/PCIDeviceService.cs
@@ -4,7 +4,7 @@ using Mosa.DeviceSystem.PCI;
 using System.Collections.Generic;
 using System.Runtime.CompilerServices;
 
-namespace Mosa.DeviceSystem
+namespace Mosa.DeviceSystem.Service
 {
 	/// <summary>
 	/// PCI Controller Service
